+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 
 **New Features and Improvements**
 
+* Tools: the command bar in the Console will now suggest completions when typing Lua expressions.
 * Runtime: meshes can now be set to cast shadows while being invisible.
 * Runtime: Android: added support for keyboard, mouse and joypad input devices.
 * Runtime: HTML5: added support for joypad input device.